+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Unreleased
 
+## Enhancements
+* Updates go-tfe client to export the instance name using `AppName()`
+
+# v1.44.0
+
+## Enhancements
+* Updates `Workspaces` to include an `AutoDestroyAt` attribute on create and update by @notchairmk and @ctrombley [#786](https://github.com/hashicorp/go-tfe/pull/786)
+* Adds `AgentsEnabled` and `PolicyToolVersion` attributes to `PolicySet` by @mrinalirao [#752](https://github.com/hashicorp/go-tfe/pull/752)
+
 # v1.43.0
 
 ## Features